rename folder to openapi_generator_tester
