+++ packages/openapi_generator_tester/analysis_options.yaml
