+++ packages/openapi_generator_tester/pubspec.yaml
@@ -1,4 +1,4 @@
-name: open_api_gen_tester
+name: openapi_generator_tester
 description: A new Flutter package project.
 version: 0.0.1
 homepage:
