+++ packages/openapi_generator_tester/test/helpers.dart
